ui/requirements.txt
streamlit==1.24.0
pandas==2.0.3
plotly==5.15.0
requests==2.31.0
pillow==9.5.0